rework rMain CodePipeline
parse sdlf-main files only once, in a dedicated lambda function
fix issue #311 where having multiple pTeamName with the same value would cause the pipeline to fail
remove usage of MaxItems when listing CloudFormation stacks as we definitely want to know about everything not just the first 30
merge domain-cicd and team-cicd lambda functions, increase timeout

frankly rMain could be improved still, but this should be more robust going forward.
