+++ sdlf-cicd/template-cicd-sdlf-pipelines.yaml
@@ -65,9 +65,9 @@ Resources:
       Parameters:
         pCicdRepository: !Ref pCicdRepository
         pMainRepository: !Ref pMainRepository
+        pMainRepositoryParserLambda: !Ref rMainRepositoryParserLambda
         pMainRepositoryDomainLambda: !Ref rMainRepositoryDomainLambda
         pMainRepositoryCrossAccountTeamLambda: !Ref rMainRepositoryCrossAccountTeamLambda
-        pMainRepositoryTeamLambda: !Ref rMainRepositoryTeamLambda
         pBuildCloudFormationPackage: !Ref rCloudFormationPackageCodeBuildProject
         pEnvironment: dev
 
@@ -83,9 +83,9 @@ Resources:
       Parameters:
         pCicdRepository: !Ref pCicdRepository
         pMainRepository: !Ref pMainRepository
+        pMainRepositoryParserLambda: !Ref rMainRepositoryParserLambda
         pMainRepositoryDomainLambda: !Ref rMainRepositoryDomainLambda
         pMainRepositoryCrossAccountTeamLambda: !Ref rMainRepositoryCrossAccountTeamLambda
-        pMainRepositoryTeamLambda: !Ref rMainRepositoryTeamLambda
         pBuildCloudFormationPackage: !Ref rCloudFormationPackageCodeBuildProject
         pEnvironment: test
 
@@ -101,9 +101,9 @@ Resources:
       Parameters:
         pCicdRepository: !Ref pCicdRepository
         pMainRepository: !Ref pMainRepository
+        pMainRepositoryParserLambda: !Ref rMainRepositoryParserLambda
         pMainRepositoryDomainLambda: !Ref rMainRepositoryDomainLambda
         pMainRepositoryCrossAccountTeamLambda: !Ref rMainRepositoryCrossAccountTeamLambda
-        pMainRepositoryTeamLambda: !Ref rMainRepositoryTeamLambda
         pBuildCloudFormationPackage: !Ref rCloudFormationPackageCodeBuildProject
         pEnvironment: prod
 
@@ -510,6 +510,26 @@ Resources:
                   - !Sub arn:${AWS::Partition}:codepipeline:${AWS::Region}:${AWS::AccountId}:sdlf-cicd-domain-*
                   - !Sub arn:${AWS::Partition}:codepipeline:${AWS::Region}:${AWS::AccountId}:sdlf-cicd-teams-*
 
+  rMainRepositoryParserLambda:
+    Type: AWS::Lambda::Function
+    Properties:
+      Description: Parses all datadomain files in the main datalake repository to get all the data domain details needed in further steps
+      Handler: lambda_function.lambda_handler
+      MemorySize: 256
+      Role: !GetAtt rMainRepositoryLambdaRole.Arn
+      KmsKeyArn: !Ref pKMSKey
+      VpcConfig: !If
+        - RunInVpc
+        - SecurityGroupIds: !Split [",", !ImportValue sdlf-cicd-prerequisites-vpc-security-groups]
+          SubnetIds: !Split [",", !ImportValue sdlf-cicd-prerequisites-vpc-subnets]
+        - !Ref "AWS::NoValue"
+      Runtime: python3.12
+      Timeout: 450
+      Environment:
+        Variables:
+          CLOUDFORMATION_ROLE: !GetAtt rMainRepositoryCloudFormationRole.Arn
+      Code: ./lambda/parser-cicd/src
+
   rMainRepositoryDomainLambda:
     Type: AWS::Lambda::Function
     Properties:
@@ -524,9 +544,12 @@ Resources:
           SubnetIds: !Split [",", !ImportValue sdlf-cicd-prerequisites-vpc-subnets]
         - !Ref "AWS::NoValue"
       Runtime: python3.12
-      Timeout: 450
+      Timeout: 720
       Environment:
         Variables:
+          AWS_PARTITION: !Ref AWS::Partition
+          DEVOPS_KMS_KEY: !Ref pKMSKey
+          MAIN_REPOSITORY_PREFIX: !Ref pMainRepositoriesPrefix
           CLOUDFORMATION_ROLE: !GetAtt rMainRepositoryCloudFormationRole.Arn
       Code: ./lambda/domain-cicd/src
 
@@ -552,29 +575,6 @@ Resources:
           ENABLE_LAMBDA_LAYER_BUILDER: !Ref pEnableLambdaLayerBuilder
           ENABLE_GLUE_JOB_DEPLOYER: !Ref pEnableGlueJobDeployer
       Code: ./lambda/crossaccountteam-cicd/src
-
-  rMainRepositoryTeamLambda:
-    Type: AWS::Lambda::Function
-    Properties:
-      Description: Creates a CodePipeline pipeline for each team in each domain-env defined in the main datalake repository
-      Handler: lambda_function.lambda_handler
-      MemorySize: 256
-      Role: !GetAtt rMainRepositoryLambdaRole.Arn
-      KmsKeyArn: !Ref pKMSKey
-      VpcConfig: !If
-        - RunInVpc
-        - SecurityGroupIds: !Split [",", !ImportValue sdlf-cicd-prerequisites-vpc-security-groups]
-          SubnetIds: !Split [",", !ImportValue sdlf-cicd-prerequisites-vpc-subnets]
-        - !Ref "AWS::NoValue"
-      Runtime: python3.12
-      Timeout: 450
-      Environment:
-        Variables:
-          AWS_PARTITION: !Ref AWS::Partition
-          DEVOPS_KMS_KEY: !Ref pKMSKey
-          MAIN_REPOSITORY_PREFIX: !Ref pMainRepositoriesPrefix
-          CLOUDFORMATION_ROLE: !GetAtt rMainRepositoryCloudFormationRole.Arn
-      Code: ./lambda/team-cicd/src
 
   ######## CODEBUILD JOBS #########
   rCloudFormationPackageCodeBuildServiceRole:
